Add aggregations support

+++ tests/App/database/migrations/2020_02_15_11000_create_books_table.php
@@ -14,7 +14,7 @@ class CreateBooksTable extends Migration
             $table->unsignedBigInteger('author_id');
             $table->string('title');
             $table->text('description');
-            $table->float('price');
+            $table->integer('price');
             $table->date('published');
             $table->softDeletes();
