+++ tests/App/elastic/migrations/2020_02_15_120000_create_books_index.php
@@ -14,7 +14,7 @@ final class CreateBooksIndex implements MigrationInterface
             $mapping->integer('author_id');
             $mapping->text('title');
             $mapping->text('description');
-            $mapping->float('price');
+            $mapping->integer('price');
             $mapping->date('published', ['format' => 'yyyy-MM-dd']);
         });
     }
